refactor(server): move fast-path fragmenter to own unit
